WikidataPageBanner option to add icons added
The extension now supports defining new icons, which can be activated by
parameters to {{PAGEBANNER...}} to render on the banner.  The icons are added as
'icons=icon1,icon2,...' as a parameter to {{PAGEBANNER}}. The icons can then be
added as a background-image property of class .oo-ui-icon-icon1{} in mediawiki
Common.css. See https://www.mediawiki.org/wiki/Extension:WikidataPageBanner
Rename getBannerUrl() to getImageUrl since it serves a more general purpose now.
Tests for icons added.

Bug: T100486
Bug: T104460
Change-Id: I36471b826e7cce9405c8d135bb3934bb0214170c