WPB fake banner cropping for small screens
On screen size below 400px, make the banners exceed 100% width so that a
cropping effect can be faked, by hiding overflow on container div and defining
focus parameters.
Adds a parameter 'origin' to {{PAGEBANNER}} which specifies focus as a
coordinate. Default focus is on centre of banner.  Example usage is
origin=0.2,0.4 where 1st value is x parameter and 2nd is y parameter. Allowed
values are between 0 and 1.
Drop 320 size from standardSizes as banner height at that size too small.

Bug: T105033
Change-Id: Ifd696117d4614389685945ad3853a473179c2c03