WikidataPageBanner redefine icons and add links
Allows editors to define the article an icon refers to. As a result, icons have
been redefined as separate parameters. Each icon parameter is of the form
icon-<iconname>=<Related Article Name>. Example icon-star=Star
Article|icon-unesco=Unesco.
This is important to let the readers know what the icons stand for and is
provided in the current Wikivoyage pagebanner template.
Corresponding tests added.

Bug: T108207
Change-Id: Ib4494910f68a1d45e470cd5b463a2ab17d3b72b6