Attempt at fixing table of contents problem
It appears $parserOutput->getTOCEnabled() is not true for pages impacted
by the bug. This avoids this check to make sure the correct result occurs.

Consider the case where the table of contents is disabled by a call
to $parserOutput->setTOCEnabled( or $outputPage->enableTOC( false );
and that is copied over to the ParserOutput object before the
WikidataPagebaner hook runs

This will lead to the table of contents not being suppressed from the  HTML
and will give an empty table of contents in the banner itself.

If a template parameter toc=yes has been added, create the table of contents
regardless. A __NOTOC__ parameter will still remove the table of contents from
the page and it will ensure that the banner takes over rendering if the table
of contents should be rendered.

I'm not sure what's leading to this race condition, but analysing the code closely
I don't see what else could be causing this problem. CirrusSearch
PageTextBuilder could be a culprit.

Bug: T121135
Change-Id: I7d17799bf02b059eda8e3ac9d52061470759f07b