Make use of …::class language feature
You might wonder how class_exists( ActualClass::class ) can work. It
does. This does not need the class to be loaded. For what PHP cares
about this is still a string. But the …::class syntax makes it possible
to find usages of a class much, much easier with the tools IDEs provide.

Change-Id: If2ae66eca264278ae503e59c96a0494f142b92a9