Allow skins to render subtitle after banner
This is a longstanding request. The banner currently overrides the handling
of the title. This means the placement of the subtitle under the banner looks
strange on certain skins. The community currently works around this by a hack
inside MediaWiki:Common.js.

To avoid this hack, this introduces a per skin config flag that allows skins to be
configured to render subtitles above the banner.

Bug: T113642
Change-Id: I802883a17080a179053e088dc1dc9468ecf8cdb7