Refine the offset margin calculation
The margin offset was not being calculated correctly.
This change moves the X and Y calculations into a common function
(because they're almost exactly the same)
and tweaks the tests to match.

Also, the fallback of 10% of the container height for when no Y offset
was provided was being used whenever a Y offset *was* provided.

Bug: T191689
Change-Id: Iff0aed3347dbfa401b5d77f146f0bba29e54b22d