Show TOC submenus on keyboard navigation
This makes the table of contents usable with a keyboard instead of a
mouse. It’s still not trivial to use using touch screen (long-pressing
the link reveals the dropdown while opening the context menu in mobile
Chrome and Firefox, but I’m not sure it’s something we can depend on,
and it’s certainly not intuitive).

This solution should work in major browsers released in the past three
years, except for legacy Edge and IE. Legacy Edge is not supported
anyway, and I couldn’t find a pure CSS solution that would work on IE11.
Since IE11 will become Grade C browser, i.e. won’t get any JavaScript,
in the foreseeable future, I don’t think a solution involving JavaScript
to support IE11 is worth the effort. This patch shouldn’t cause any
feature degradation in IE11, so we simply stay at the status quo there.

However, it does worsen the situation described in T301088, as now more
links are available before reaching the jump links. I think it still
makes sense to fix this issue, though.

Bug: T168506
Change-Id: I7a919efeabe1ecc53e62d1138114079ee03c5a79