Replace isset() with string or null check
isset() should only be used to suppress errors, not for null check.
When the variable is always defined, there is no need to use isset.

Replace the check for $iconname with empty string as substr() never
returns null (it returns false in older php versions)

Found by a new phan plugin (2efea9f989)
https://www.mediawiki.org/wiki/Manual:Coding_conventions/PHP#isset

Change-Id: I89a22fdff065539f6f2f17d0ab41808e906a706a